backend-app/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
packaging==26.0
psycopg2==2.9.11
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
